3.3.8a6 snapshot
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk/x3270@127 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ Common/ft_dft.c
@@ -414,7 +414,7 @@ dft_get_request(void)
 
 	if (total_read) {
 		trace_ds("> WriteStructuredField FileTransferData Data(rec=%lu) %d bytes\n",
-		    recnum, total_read);
+		    recnum, (int)total_read);
 		SET16(obptr, TR_GET_REPLY);
 		SET16(obptr, TR_RECNUM_HDR);
 		SET32(obptr, recnum);
